// ConnectionRestApiClients/clients/typescript-node/src/model/baseTemplateConversion.ts
/**
 * Connection Rest API 1.0
 * IDEA StatiCa Connection API, used for the automated design and calculation of steel connections.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: info@ideastatica.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class BaseTemplateConversion {
    'originalValue'?: string | null;
    'originalTemplateId'?: string | null;
    'newValue'?: string | null;
    'description'?: string | null;
    'newTemplateId'?: string | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "originalValue",
            "baseName": "originalValue",
            "type": "string"
        },
        {
            "name": "originalTemplateId",
            "baseName": "originalTemplateId",
            "type": "string"
        },
        {
            "name": "newValue",
            "baseName": "newValue",
            "type": "string"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "newTemplateId",
            "baseName": "newTemplateId",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return BaseTemplateConversion.attributeTypeMap;
    }
}