// ConnectionRestApiClients/clients/typescript-node/src/model/boltGrid.ts
/**
 * Connection Rest API 1.0
 * API for designing steel connections
 *
 * The version of the OpenAPI document: 1.0
 * Contact: info@ideastatica.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { BoltShearType } from './boltShearType';
import { Point3D } from './point3D';
import { ReferenceElement } from './referenceElement';
import { Vector3D } from './vector3D';

export class BoltGrid {
    /**
    * Indicates, whether a shear plane is in the thread of a bolt.
    */
    'shearInThread'?: boolean;
    'boltInteraction'?: BoltShearType;
    'boltAssembly'?: ReferenceElement;
    'origin'?: Point3D;
    'axisX'?: Vector3D;
    'axisY'?: Vector3D;
    'axisZ'?: Vector3D;
    /**
    * Positions of holes in the local coordinate system of the grid
    */
    'positions'?: Array<Point3D> | null;
    /**
    * List of the connected parts
    */
    'connectedParts'?: Array<ReferenceElement> | null;
    /**
    * Name
    */
    'name'?: string | null;
    /**
    * Length
    */
    'length'?: number;
    /**
    * Element Id
    */
    'id'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "shearInThread",
            "baseName": "shearInThread",
            "type": "boolean"
        },
        {
            "name": "boltInteraction",
            "baseName": "boltInteraction",
            "type": "BoltShearType"
        },
        {
            "name": "boltAssembly",
            "baseName": "boltAssembly",
            "type": "ReferenceElement"
        },
        {
            "name": "origin",
            "baseName": "origin",
            "type": "Point3D"
        },
        {
            "name": "axisX",
            "baseName": "axisX",
            "type": "Vector3D"
        },
        {
            "name": "axisY",
            "baseName": "axisY",
            "type": "Vector3D"
        },
        {
            "name": "axisZ",
            "baseName": "axisZ",
            "type": "Vector3D"
        },
        {
            "name": "positions",
            "baseName": "positions",
            "type": "Array<Point3D>"
        },
        {
            "name": "connectedParts",
            "baseName": "connectedParts",
            "type": "Array<ReferenceElement>"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "length",
            "baseName": "length",
            "type": "number"
        },
        {
            "name": "id",
            "baseName": "id",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return BoltGrid.attributeTypeMap;
    }
}

export namespace BoltGrid {
}