// ConnectionRestApiClients/clients/typescript-node/src/model/checkResConcreteBlock.ts
/**
 * Connection Rest API 1.0
 * IDEA StatiCa Connection API, used for the automated design and calculation of steel connections.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: info@ideastatica.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* Check value for Concrete Block
*/
export class CheckResConcreteBlock {
    /**
    * Name of Concrete Block
    */
    'name'?: string | null;
    /**
    * Unity Check
    */
    'unityCheck'?: number;
    /**
    * Status of the Check
    */
    'checkStatus'?: boolean;
    /**
    * Id of Load Case
    */
    'loadCaseId'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "unityCheck",
            "baseName": "unityCheck",
            "type": "number"
        },
        {
            "name": "checkStatus",
            "baseName": "checkStatus",
            "type": "boolean"
        },
        {
            "name": "loadCaseId",
            "baseName": "loadCaseId",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return CheckResConcreteBlock.attributeTypeMap;
    }
}