// ConnectionRestApiClients/clients/typescript-node/src/model/checkResSummary.ts
/**
 * Connection Rest API 1.0
 * API for designing steel connections
 *
 * The version of the OpenAPI document: 1.0
 * Contact: info@ideastatica.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* Check summary
*/
export class CheckResSummary {
    /**
    * Check value
    */
    'checkValue'?: number;
    /**
    * Status of check
    */
    'checkStatus'?: boolean;
    /**
    * Id of Load Case
    */
    'loadCaseId'?: number;
    /**
    * Name
    */
    'name'?: string | null;
    /**
    * Detail message about overall check
    */
    'unityCheckMessage'?: string | null;
    /**
    * Whether the check was calculated or not.  If true, the check was not calculated and IdeaRS.OpenModel.Connection.CheckResSummary.CheckValue should be ignored, otherwise false.
    */
    'skipped'?: boolean;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "checkValue",
            "baseName": "checkValue",
            "type": "number"
        },
        {
            "name": "checkStatus",
            "baseName": "checkStatus",
            "type": "boolean"
        },
        {
            "name": "loadCaseId",
            "baseName": "loadCaseId",
            "type": "number"
        },
        {
            "name": "name",
            "baseName": "name",
            "type": "string"
        },
        {
            "name": "unityCheckMessage",
            "baseName": "unityCheckMessage",
            "type": "string"
        },
        {
            "name": "skipped",
            "baseName": "skipped",
            "type": "boolean"
        }    ];

    static getAttributeTypeMap() {
        return CheckResSummary.attributeTypeMap;
    }
}