// ConnectionRestApiClients/clients/typescript-node/src/model/conCalculationParameter.ts
/**
 * Connection Rest API 1.0
 * API for designing steel connections
 *
 * The version of the OpenAPI document: 1.0
 * Contact: info@ideastatica.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ConAnalysisTypeEnum } from './conAnalysisTypeEnum';

export class ConCalculationParameter {
    'connectionIds'?: Array<number> | null;
    'analysisType'?: ConAnalysisTypeEnum;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "connectionIds",
            "baseName": "connectionIds",
            "type": "Array<number>"
        },
        {
            "name": "analysisType",
            "baseName": "analysisType",
            "type": "ConAnalysisTypeEnum"
        }    ];

    static getAttributeTypeMap() {
        return ConCalculationParameter.attributeTypeMap;
    }
}

export namespace ConCalculationParameter {
}